requirements.txt
numpy
scipy
imageio
spatialmath-python
matplotlib
